Improved option 5 results.

+++ src/main/Calculator.java
@@ -4,9 +4,9 @@ import java.util.Arrays;
 
 public class Calculator {
 
-	private int add;
-	private int sub;
-	private int mul;
+	private long add;
+	private long sub;
+	private long mul;
 
 	public Calculator() {
 		add = 0;
@@ -156,7 +156,7 @@ public class Calculator {
 		if (a.getBase() != b.getBase()) {
 			throw new IllegalArgumentException("Both numbers have to be represented in the same base.");
 		}
-		
+
 		int base = a.getBase();
 		int length = Math.max(a.getLength(), b.getLength());
 		if (a.getLength() < b.getLength()) {
@@ -168,15 +168,16 @@ public class Calculator {
 		int[][] subresult = new int[a.getLength()][2 * length];
 		int prelres = 0;
 
-		//Find the sign of the result
+		// Find the sign of the result
 		boolean resultSign = a.isNegative() ^ b.isNegative();
 
-		if (a.isNegative()) 
+		if (a.isNegative())
 			a = a.flipSign();
 		if (b.isNegative())
 			b = b.flipSign();
 
-		//The primary school multiplication (multiplying digit by digit and putting each subresult in a matrix)
+		// The primary school multiplication (multiplying digit by digit and
+		// putting each subresult in a matrix)
 		int carrier = 0;
 		for (int i = 0; i < b.getLength(); i++) {
 			carrier = 0;
@@ -199,7 +200,7 @@ public class Calculator {
 			}
 		}
 
-		//The addition column by column for the final result
+		// The addition column by column for the final result
 		for (int i = 2 * length - 1; i >= 0; i--) {
 			for (int j = 0; j < b.getLength(); j++) {
 				if (subresult[j][i] != 0) {
@@ -210,7 +211,8 @@ public class Calculator {
 		}
 		carrier = 0;
 
-		//Making sure that each "digit" of the result isn't equal to or greater than the base (the carries for the result)
+		// Making sure that each "digit" of the result isn't equal to or greater
+		// than the base (the carries for the result)
 		for (int i = 0; i < 2 * length; i++) {
 			if (carrier != 0) {
 				transfer[i] = transfer[i] + carrier;
@@ -223,8 +225,8 @@ public class Calculator {
 				carrier = 0;
 			}
 		}
-		
-		//The final result (represented in the proper base)
+
+		// The final result (represented in the proper base)
 		return new Number(base, transfer, resultSign);
 	}
 
@@ -320,15 +322,15 @@ public class Calculator {
 		mul = 0;
 	}
 
-	public int getAdd() {
+	public long getAdd() {
 		return add;
 	}
 
-	public int getSub() {
+	public long getSub() {
 		return sub;
 	}
 
-	public int getMul() {
+	public long getMul() {
 		return mul;
 	}
